requirements.txt
requests
beautifulsoup4
faiss-cpu
sentence-transformers
langchain
langchain-community
langchain-text-splitters

transformers
python-dotenv
streamlit
